+++ src/main/java/top/sywyar/pixivdownload/common/AppInfo.java
@@ -1,0 +1,35 @@
+package top.sywyar.pixivdownload.common;
+
+import lombok.experimental.UtilityClass;
+
+/**
+ * 应用级固定信息。
+ */
+@UtilityClass
+public class AppInfo {
+
+    public static final String NAME = "PixivDownloader";
+    public static final String LEGACY_ARTIFACT_NAME = "PixivDownload";
+    public static final String EXECUTABLE_NAME = LEGACY_ARTIFACT_NAME + ".exe";
+    public static final String SHORTCUT_NAME = LEGACY_ARTIFACT_NAME + ".lnk";
+    public static final String HIDDEN_DIRECTORY_NAME = ".pixivdownload";
+
+    public static final String GITHUB_REPOSITORY = "Sywyar/" + NAME;
+    public static final String GITHUB_URL = "https://github.com/" + GITHUB_REPOSITORY;
+    public static final String RELEASES_URL = GITHUB_URL + "/releases";
+    public static final String WIKI_URL = GITHUB_URL + "/wiki";
+    public static final String LICENSE_URL = GITHUB_URL + "/blob/master/LICENSE";
+    public static final String LATEST_RELEASE_DOWNLOAD_URL = RELEASES_URL + "/latest/download";
+    public static final String NIGHTLY_RELEASE_DOWNLOAD_URL = RELEASES_URL + "/download/nightly";
+
+    public static final String MAVEN_POM_PROPERTIES_RESOURCE =
+            "/META-INF/maven/top.sywyar.lovepopup/" + LEGACY_ARTIFACT_NAME + "/pom.properties";
+
+    public static String userAgent(String component) {
+        return NAME + "/" + component;
+    }
+
+    public static String dashedUserAgent(String component) {
+        return NAME + "-" + component;
+    }
+}